+++ react/components/EmailAndPassword.js
@@ -1,7 +1,9 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
-import { Input, Button, IconArrowBack } from 'vtex.styleguide'
 import { injectIntl, intlShape } from 'react-intl'
+
+import { Input, Button } from 'vtex.styleguide'
+import { AuthState, AuthService } from 'vtex.auth'
 
 import { translate } from '../utils/translate'
 import { isValidEmail, isValidPassword } from '../utils/format-check'
@@ -9,8 +11,8 @@ import { steps } from '../utils/steps'
 import Form from './Form'
 import FormError from './FormError'
 import PasswordInput from './PasswordInput'
+import GoBackButton from './GoBackButton'
 
-import { AuthState, AuthService } from 'vtex.auth'
 
 /** EmailAndPasswordLogin component. */
 class EmailAndPassword extends Component {
@@ -170,22 +172,10 @@ class EmailAndPassword extends Component {
         footer={
           <Fragment>
             {showBackButton && (
-              <div className="vtex-login__back-button">
-                <Button
-                  variation="tertiary"
-                  size="small"
-                  onClick={() =>
-                    onStateChange({ step: previous, password: '' })
-                  }
-                >
-                  <span className="vtex-login__back-icon c-link">
-                    <IconArrowBack size={10} color="currentColor" />
-                  </span>
-                  <span className="t-small ml2">
-                    {translate('login.goBack', intl)}
-                  </span>
-                </Button>
-              </div>
+              <GoBackButton
+                onStateChange={onStateChange}
+                changeTab={{ step: previous, password: '' }}
+              />
             )}
             <div className="vtex-login__send-button">
               <AuthService.LoginWithPassword