+++ react/components/LoginComponent.js
@@ -1,15 +1,18 @@
 import React, { Component, Fragment } from 'react'
 import { Link } from 'render'
-import { Button } from 'vtex.styleguide'
 import OutsideClickHandler from 'react-outside-click-handler'
+import classNames from 'classnames'
+
+import { ButtonWithIcon } from 'vtex.styleguide'
+import Icon from 'vtex.use-svg/Icon'
 
 import LoginContent from '../LoginContent'
-import Icon from 'vtex.use-svg/Icon'
 import { truncateString } from '../utils/format-string'
 import { translate } from '../utils/translate'
 import { LoginPropTypes } from '../propTypes'
 import { getProfile } from '../utils/profile'
 
+const profileIcon = iconSize => (<Icon id="hpa-profile" size={iconSize} />)
 export default class LoginComponent extends Component {
   static propTypes = LoginPropTypes
 
@@ -25,7 +28,6 @@ export default class LoginComponent extends Component {
     const {
       iconSize,
       iconLabel,
-      iconClasses,
       labelClasses,
       intl,
       renderIconAsLink,
@@ -37,9 +39,11 @@ export default class LoginComponent extends Component {
 
     const iconContent = (
       <Fragment>
-        <div className={`${iconClasses}`}>
-          <Icon id="hpa-profile" size={iconSize} />
-        </div>
+        {renderIconAsLink &&
+          <div className="flex items-center">
+            <Icon id="hpa-profile" size={iconSize} />
+          </div>
+        }
         {profile ? (
           <span className={`vtex-login__profile order-1 f6 pl4 ${labelClasses} dn-m db-l`}>
             {translate('login.hello', intl)}{' '}
@@ -64,7 +68,7 @@ export default class LoginComponent extends Component {
     }
 
     return (
-      <Button variation="tertiary" icon onClick={onProfileIconClick}>
+      <ButtonWithIcon variation="tertiary" icon={profileIcon(iconSize)} iconPosition="left" onClick={onProfileIconClick}>
         <div
           className="flex pv2 items-center"
           ref={e => {
@@ -73,7 +77,7 @@ export default class LoginComponent extends Component {
         >
           {iconContent}
         </div>
-      </Button>
+      </ButtonWithIcon>
     )
   }
 
@@ -91,9 +95,10 @@ export default class LoginComponent extends Component {
           {this.renderIcon()}
           <OutsideClickHandler onOutsideClick={onOutSideBoxClick}>
             <div
-              className={`vtex-login__box absolute z-max ${
-                isBoxOpen ? 'flex' : 'dn'
-                }`}
+              className={classNames('vtex-login__box absolute z-max', {
+                'flex': isBoxOpen,
+                'dn': !isBoxOpen,
+              })}
               style={boxPositionStyle}
             >
               <div className="vtex-login__arrow-up absolute top-0 right-0 shadow-3 bg-base" />